refactor(scripts): ライセンス生成ロジックをproduct/web間で共通化
apps/product用 scripts/generate-licenses.ts と apps/web用
apps/web/scripts/generate-licenses.ts が同じpnpm licenses list --json
ベースのロジックをほぼ丸ごと重複して持っていた状態(#1654のCodex review
指摘対応で発覚)を解消。scripts/lib/license-generator.tsに共通ロジックを
抽出し、両エントリスクリプトはpackage名・出力先・ラベルだけを渡す薄い
wrapperにした。

あわせてapps/webをCIのlicense:credits:checkに組み込み。従来は
apps/product分しかcheckされておらず、apps/webのoss-credits.jsonの
鮮度がCIで一切検証されていなかった(root package.jsonのlicense:credits:check
/ generate-licensesを両パッケージ横断に変更)。

apps/product側の出力(oss-credits.json / THIRD_PARTY_NOTICES.txt)は
リファクタ前後で差分なしを確認済み。apps/web側はfooterのタイムスタンプ
表記をproduct同様の固定文言に統一したため1行差分が出ている。

テストは scripts/__tests__/generate-licenses.test.ts を
scripts/lib/__tests__/license-generator.test.ts に移設。
