+++ apps/web/package.json
@@ -50,6 +50,7 @@
     "license:check": "pnpm exec license-checker --production --excludePackages \"@dayopt/web@$(node -p 'require(\\\"./package.json\\\").version')\" --onlyAllow \"$(cat .licensrc.json | jq -r .onlyAllow)\"",
     "license:check-risks": "tsx scripts/check-license-risks.ts",
     "license:audit": "pnpm exec license-checker --production --summary",
+    "license:credits:check": "tsx scripts/generate-licenses.ts --check",
     "generate-licenses": "tsx scripts/generate-licenses.ts",
     "validate:content": "node scripts/validate-content.js",
     "generate:search-index": "tsx scripts/generate-search-index.ts"
