+++ apps/web/public/THIRD_PARTY_NOTICES.txt
@@ -99,4 +99,4 @@ available under the Apache 2.0 license (https://github.com/puppeteer/puppeteer/b
 For a complete list of all third-party software and their licenses,
 please visit: /legal/oss-credits
 
-Generated: 2026-07-20T23:18:55.238Z
+Generated by apps/web/scripts/generate-licenses.ts.
