+++ package.json
@@ -50,8 +50,8 @@
     "license:check": "pnpm --filter @dayopt/product license:check",
     "license:check-risks": "pnpm --filter @dayopt/product license:check-risks",
     "license:audit": "pnpm --filter @dayopt/product license:audit",
-    "license:credits:check": "pnpm --filter @dayopt/product license:credits:check",
-    "generate-licenses": "pnpm --filter @dayopt/product generate-licenses",
+    "license:credits:check": "pnpm --filter @dayopt/product license:credits:check && pnpm --filter @dayopt/web license:credits:check",
+    "generate-licenses": "pnpm --filter @dayopt/product generate-licenses && pnpm --filter @dayopt/web generate-licenses",
     "perf:lighthouse": "pnpm --filter @dayopt/product perf:lighthouse",
     "quality:full": "pnpm --filter @dayopt/product quality:full",
     "quality:deadcode": "pnpm --filter @dayopt/product quality:deadcode",
