+++ scripts/lib/__tests__/license-generator.test.ts
@@ -6,7 +6,7 @@ import {
   getProductionDependencyFilter,
   normalizeGeneratedFileForCheck,
   parsePnpmLicenses,
-} from '../generate-licenses';
+} from '../license-generator';
 
 describe('getProductionDependencyFilter', () => {
   it('includes workspace package dependencies in the pnpm license scan', () => {
@@ -105,10 +105,16 @@ describe('parsePnpmLicenses', () => {
 
 describe('generateNoticesFile', () => {
   it('does not include wall-clock timestamps', () => {
-    const content = generateNoticesFile([]);
+    const content = generateNoticesFile([], 'scripts/generate-licenses.ts');
 
     expect(content).toContain('Generated by scripts/generate-licenses.ts.');
     expect(content).not.toMatch(/Generated: \d{4}-\d{2}-\d{2}/);
+  });
+
+  it('labels the generating script per caller', () => {
+    const content = generateNoticesFile([], 'apps/web/scripts/generate-licenses.ts');
+
+    expect(content).toContain('Generated by apps/web/scripts/generate-licenses.ts.');
   });
 });
